i added give us 8 and u get yes

+++ CS01-6.py
@@ -1,3 +1,3 @@
-A = int(input("Give us a number: "))
+A = float(input("Give us a number: "))
 if (A<=10):
     print("Yes")
